Renamed readme.txt to configure_db.md and placed it into a new docs directory in the root of the directory. Also found a spelling mistake in the persistence package so renamed that and refactored code to match.
