+++ src/main/java/persistence/ClientCredentials.java
@@ -1,4 +1,4 @@
-package persistance;
+package persistence;
 
 import javax.persistence.*;
