+++ src/main/resources/META-INF/persistence.xml
@@ -5,8 +5,8 @@
 
     <persistence-unit name="fitbitPU">
         <provider>org.hibernate.jpa.HibernatePersistenceProvider</provider>
-        <class>persistance.TokenMap</class>
-        <class>persistance.ClientCredentials</class>
+        <class>persistence.TokenMap</class>
+        <class>persistence.ClientCredentials</class>
         <properties>
             <property name="hibernate.connection.url" value=""/>
             <property name="hibernate.connection.driver_class" value=""/>
