+++ src/test/java/TestPersistance_Local.java
@@ -1,7 +1,7 @@
 import org.junit.Assert;
 import org.junit.Test;
 import org.junit.Assume;
-import persistance.StorageManager;
+import persistence.StorageManager;
 
 public class TestPersistance_Local {
